+++ src/main/java/mod/bluestaggo/modernerbeta/util/BlockStates.java
@@ -15,7 +15,7 @@ public class BlockStates {
     public static final BlockState SANDSTONE = Blocks.SANDSTONE.getDefaultState();
     public static final BlockState RED_SAND = Blocks.RED_SAND.getDefaultState();
     public static final BlockState RED_SANDSTONE = Blocks.RED_SANDSTONE.getDefaultState();
-    public static final BlockState ORANGE_TERRACOTTA = Blocks.ORANGE_TERRACOTTA.getDefaultState();
+    public static final BlockState TERRACOTTA = Blocks.TERRACOTTA.getDefaultState();
     
     public static final BlockState STONE = Blocks.STONE.getDefaultState();
     public static final BlockState AIR = Blocks.AIR.getDefaultState();