+++ src/main/java/mod/bluestaggo/modernerbeta/world/biome/provider/fractal/predicates/NeighborComparisonPredicate.java
@@ -39,8 +39,6 @@ public abstract class NeighborComparisonPredicate extends BiomePredicate {
         int[] xCoords = this.diagonal ? DIAGONAL_NEIGHBOR_X_COORDS : NEIGHBOR_X_COORDS;
         int[] zCoords = this.diagonal ? DIAGONAL_NEIGHBOR_Z_COORDS : NEIGHBOR_Z_COORDS;
 
-        this.prepareMatch(biome, layer, randomSupplier, x, z);
-
         int matches = 0;
         for (int i = 0; i < 4; i++) {
             int nx = x + xCoords[i];
@@ -53,9 +51,6 @@ public abstract class NeighborComparisonPredicate extends BiomePredicate {
         }
 
         return false;
-    }
-
-    protected void prepareMatch(ExtendedBiomeId biome, Layer layer, Supplier<LayerRandom> randomSupplier, int x, int z) {
     }
 
     protected abstract boolean neighborMatches(ExtendedBiomeId centre, ExtendedBiomeId neighbor, Layer layer, Supplier<LayerRandom> randomSupplier, int x, int z, int nx, int nz);