+++ src/main/java/mod/bluestaggo/modernerbeta/world/chunk/provider/ChunkProviderBeta.java
@@ -81,7 +81,7 @@ public class ChunkProviderBeta extends ChunkProviderNoise {
         BlockPos.Mutable pos = new BlockPos.Mutable();
         
         AquiferSampler aquiferSampler = this.getAquiferSampler(chunk, noiseConfig);
-        ChunkHeightmap heightmapChunk = this.getChunkHeightmap(chunkX, chunkZ);
+        ChunkHeightmap heightmapChunk = this.hasNoisePostProcessor() ? this.getChunkHeightmap(chunkX, chunkZ) : null;
         SimpleNoisePos noisePos = new SimpleNoisePos();
         
         double[] sandNoise = beachOctaveNoise.sampleBeta(
@@ -105,7 +105,7 @@ public class ChunkProviderBeta extends ChunkProviderNoise {
                 int x = startX + localX;
                 int z = startZ + localZ;
                 int surfaceTopY = chunk.getHeightmap(Heightmap.Type.OCEAN_FLOOR_WG).get(localX, localZ) - 1;
-                int surfaceMinY = (this.hasNoisePostProcessor()) ? 
+                int surfaceMinY = heightmapChunk != null ?
                     heightmapChunk.getHeight(x, z, ChunkHeightmap.Type.SURFACE_FLOOR) - 8 : 
                     this.worldMinY;
                 
@@ -223,6 +223,7 @@ public class ChunkProviderBeta extends ChunkProviderNoise {
         int startZ = chunk.getPos().getStartZ();
 
         Random rand = this.createSurfaceRandom(chunkX, chunkZ);
+        ChunkHeightmap heightmapChunk = this.hasNoisePostProcessor() ? this.getChunkHeightmap(chunkX, chunkZ) : null;
         BlockPos.Mutable pos = new BlockPos.Mutable();
 
         double[] sandNoise = beachOctaveNoise.sampleBeta(
@@ -244,9 +245,12 @@ public class ChunkProviderBeta extends ChunkProviderNoise {
         for (int localZ = 0; localZ < 16; localZ++) {
             for (int localX = 0; localX < 16; localX++) {
                 pos.set(localX, 0, localZ);
+
                 int x = startX + localX;
                 int z = startZ + localZ;
-                int surfaceTopY = chunk.getHeightmap(Heightmap.Type.OCEAN_FLOOR_WG).get(localX, localZ) - 1;
+                int surfaceTopY = heightmapChunk != null ?
+                    heightmapChunk.getHeight(x, z, ChunkHeightmap.Type.SURFACE_FLOOR) :
+                    chunk.getHeightmap(Heightmap.Type.WORLD_SURFACE_WG).get(localX, localZ) - 1;
 
                 boolean genSandBeach = sandNoise[localZ + localX * 16] + rand.nextDouble() * 0.2D > 0.0D;
                 boolean genGravelBeach = gravelNoise[localZ + localX * 16] + rand.nextDouble() * 0.2D > 3D;